message listener cleanup
